requirements.txt
Django>=3.0,<4.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
django-environ==0.10.0
djangorestframework==3.11.2